refactor(livechat): split Chat route into content and footer components
Extract the message list into ChatContent and the composer/options into
ChatFooter, and move the lifecycle hook into its own useChatEffects module.
Chat now owns the shared inputRef/notifyEmojiSelectRef, the emoji-picker
state, and the upload flow, passing handlers down to the two subcomponents.
